+++ ckeditor5_plugins/jumpmenu/src/index.js
@@ -1,0 +1,3 @@
+import JumpMenu from './jumpmenu';
+
+export default { JumpMenu };